requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
pandas==2.0.3
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-docx==0.8.11
openpyxl==3.1.2
